sloution3

+++ src/main/java/com/driver/services/BlogService.java
@@ -28,6 +28,9 @@ public class BlogService {
         blog.setTitle(title);
 
         User user = userRepository1.findById(userId).get();
+        if(user == null){
+            return null;
+        }
         List<Blog> blogList = user.getBlogList();
         blogList.add(blog);
         user.setBlogList(blogList);
@@ -40,6 +43,9 @@ public class BlogService {
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
         Blog blog = blogRepository1.findById(blogId).get();
+        if(blog == null){
+            return;
+        }
         blogRepository1.deleteById(blogId);
         User user = blog.getUser();
         List<Blog> blogList = user.getBlogList();
